Component Header changed to NavBar. Style changes applied - MariaD 12.20.24
