+++ package.json
@@ -22,7 +22,7 @@
     "dayjs": "^1.11.13",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^7.0.1",
+    "react-router-dom": "^7.0.2",
     "styled-components": "^6.1.13"
   },
   "devDependencies": {
