+++ src/components/Footer.jsx
@@ -15,16 +15,20 @@ const Footer = () => {
         display: "flex",
         justifyContent: "space-between",
         alignItems: "center",
+        position: "fixed",
+        bottom: 0,
+        left: 0,
+        width: "100%",
       }}
     >
-      <Typography variant="body1">© 2024 App Name</Typography>
+      <Typography variant="body2">© 2024 App Name</Typography>
       <Box
         sx={{
           display: "flex",
           alignItems: "center", // Aligns "Follow Us" and icons vertically
         }}
       >
-        <Typography variant="body1" sx={{ marginRight: 1 }}>
+        <Typography variant="body2" sx={{ marginRight: 1 }}>
           Follow Us
         </Typography>
         <a
